refactoring to traits
