+++ src/Behaviors/ConfiguresCashier.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MichaelRubel\StripeIntegration\Behaviors;
+
+use Money\Currency;
+
+trait ConfiguresCashier
+{
+    /**
+     * Update the Cashier currency.
+     *
+     * @param  Currency  $currency
+     *
+     * @return void
+     */
+    public function setCashierCurrency(Currency $currency): void
+    {
+        config([
+            'cashier.currency' => $currency->getCode(),
+        ]);
+    }
+}
